+++ app/src/main/java/petstone/project/animalisland/activity/MainActivity.java
@@ -1,21 +1,17 @@
 package petstone.project.animalisland.activity;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 
 import android.os.Bundle;
 import android.view.MenuItem;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
-import petstone.project.animalisland.Free_Rehome;
 import petstone.project.animalisland.R;
-import petstone.project.animalisland.Sell_Rehome;
+import petstone.project.animalisland.component.ChatComponent;
 import petstone.project.animalisland.component.MypageComponent;
+import petstone.project.animalisland.component.PetFriendComponent;
 import petstone.project.animalisland.component.RehomeComponent;
 
 //메인
@@ -24,8 +20,8 @@ public class MainActivity extends AppCompatActivity {
     //선언부
     BottomNavigationView bottomNavi;
     RehomeComponent homeFrag;
-    PetFriend petFrag;
-    ChatActivity chatFrag;
+    PetFriendComponent petFrag;
+    ChatComponent chatFrag;
     MypageComponent myFrag;
 
 
@@ -39,9 +35,9 @@ public class MainActivity extends AppCompatActivity {
 
 
         bottomNavi = findViewById(R.id.bottom_tab);
-        chatFrag = new ChatActivity();
+        chatFrag = new ChatComponent();
         homeFrag = new RehomeComponent();
-        petFrag = new PetFriend();
+        petFrag = new PetFriendComponent();
         myFrag = new MypageComponent();
 
         getSupportFragmentManager().beginTransaction().replace(R.id.main_frame,homeFrag).commitAllowingStateLoss();
@@ -57,7 +53,7 @@ public class MainActivity extends AppCompatActivity {
                         getSupportFragmentManager().beginTransaction().replace(R.id.main_frame, petFrag).commitAllowingStateLoss();
                         return true;
                     case R.id.bottom_chat:
-                        //getSupportFragmentManager().beginTransaction().replace(R.id.main_frame, chatFrag).commitAllowingStateLoss();
+                        getSupportFragmentManager().beginTransaction().replace(R.id.main_frame, chatFrag).commitAllowingStateLoss();
                         return true;
                     case R.id.bottom_mypage:
                         getSupportFragmentManager().beginTransaction().replace(R.id.main_frame, myFrag).commitAllowingStateLoss();
