+++ app/src/main/java/petstone/project/animalisland/activity/SelectFreeRehomeActivity.java
@@ -1,19 +1,22 @@
-package petstone.project.animalisland;
+package petstone.project.animalisland.activity;
 
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageView;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class Select_Sell_Rehome extends AppCompatActivity {
+import petstone.project.animalisland.R;
+
+public class SelectFreeRehomeActivity extends AppCompatActivity {
 
     ImageView back;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.select_sell_rehome);
+        setContentView(R.layout.select_free_rehome);
 
         back = findViewById(R.id.back);
