+++ app/src/main/java/petstone/project/animalisland/other/FreeRecycleAdapter.java
@@ -1,4 +1,4 @@
-package petstone.project.animalisland;
+package petstone.project.animalisland.other;
 
 import android.content.Context;
 import android.content.Intent;
@@ -12,33 +12,36 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 
-public class SellRecycleAdapter extends RecyclerView.Adapter<SellRecycleAdapter.ViewHolder> {
+import petstone.project.animalisland.R;
+import petstone.project.animalisland.activity.SelectFreeRehomeActivity;
 
-    private ArrayList<SellRehomeList> mData = null ;
-    private Context s_context;
+public class FreeRecycleAdapter extends RecyclerView.Adapter<FreeRecycleAdapter.ViewHolder> {
+
+    private ArrayList<FreeRehomeList> mData = null ;
+    private Context context;
 
     // 생성자에서 데이터 리스트 객체를 전달받음.
-    SellRecycleAdapter(ArrayList<SellRehomeList> list) {
+    FreeRecycleAdapter(ArrayList<FreeRehomeList> list) {
         mData = list ;
     }
 
     // onCreateViewHolder() - 아이템 뷰를 위한 뷰홀더 객체 생성하여 리턴.
     @Override
-    public SellRecycleAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        s_context = parent.getContext() ;
-        LayoutInflater inflater = (LayoutInflater) s_context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) ;
+    public FreeRecycleAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+         context = parent.getContext() ;
+        LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) ;
 
-        View view = inflater.inflate(R.layout.sell_rehome_list, parent, false) ;
-        SellRecycleAdapter.ViewHolder vh = new SellRecycleAdapter.ViewHolder(view) ;
+        View view = inflater.inflate(R.layout.free_rehome_list, parent, false) ;
+        FreeRecycleAdapter.ViewHolder vh = new FreeRecycleAdapter.ViewHolder(view) ;
 
         return vh ;
     }
 
     // onBindViewHolder() - position에 해당하는 데이터를 뷰홀더의 아이템뷰에 표시.
     @Override
-    public void onBindViewHolder(SellRecycleAdapter.ViewHolder holder, int position) {
+    public void onBindViewHolder(FreeRecycleAdapter.ViewHolder holder, int position) {
 
-        SellRehomeList item = mData.get(position) ;
+        FreeRehomeList item = mData.get(position) ;
 
         holder.main_img.setImageDrawable(item.getImg()) ;
         holder.gender.setImageDrawable(item.getGender()) ;
@@ -46,7 +49,6 @@ public class SellRecycleAdapter extends RecyclerView.Adapter<SellRecycleAdapter.
         holder.age.setText(item.getAge()) ;
         holder.local.setText(item.getLocal()) ;
         holder.date.setText(item.getDate()) ;
-        holder.price.setText(item.getPrice());
         holder.nickname.setText(item.getNickname()) ;
     }
 
@@ -59,20 +61,18 @@ public class SellRecycleAdapter extends RecyclerView.Adapter<SellRecycleAdapter.
 
     public class ViewHolder extends RecyclerView.ViewHolder {
         ImageView main_img, gender ;
-        TextView breed, age, local, date, price, nickname ;
+        TextView breed, age, local, date, nickname ;
 
         ViewHolder(View itemView) {
             super(itemView) ;
 
             main_img = itemView.findViewById(R.id.main_img) ;
-            gender = itemView.findViewById(R.id.sell_gender) ;
-            breed = itemView.findViewById(R.id.sell_breed) ;
-            age = itemView.findViewById(R.id.sell_age) ;
-            local = itemView.findViewById(R.id.sell_local) ;
-            date = itemView.findViewById(R.id.sell_date) ;
-            price = itemView.findViewById(R.id.sell_price);
-
-            nickname = itemView.findViewById(R.id.sell_nickname) ;
+            gender = itemView.findViewById(R.id.free_gender) ;
+            breed = itemView.findViewById(R.id.free_breed) ;
+            age = itemView.findViewById(R.id.free_age) ;
+            local = itemView.findViewById(R.id.free_local) ;
+            date = itemView.findViewById(R.id.free_date) ;
+            nickname = itemView.findViewById(R.id.free_nickname) ;
 
             itemView.setClickable(true);
 
@@ -81,8 +81,8 @@ public class SellRecycleAdapter extends RecyclerView.Adapter<SellRecycleAdapter.
                 public void onClick(View view) {
                     int pos = getAdapterPosition();
                     if(pos != RecyclerView.NO_POSITION){
-                        Intent intent = new Intent(s_context, Select_Sell_Rehome.class);
-                        s_context.startActivity(intent);
+                        Intent intent = new Intent(context, SelectFreeRehomeActivity.class);
+                        context.startActivity(intent);
                     }
                 }
             });
