+++ app/src/main/java/petstone/project/animalisland/other/MypageCustomListAdapter.java
@@ -1,4 +1,4 @@
-package petstone.project.animalisland.component;
+package petstone.project.animalisland.other;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
