+++ app/src/main/java/petstone/project/animalisland/other/MypageNewsDialog.java
@@ -1,4 +1,4 @@
-package petstone.project.animalisland.component;
+package petstone.project.animalisland.other;
 
 import android.app.Dialog;
 import android.content.Context;
@@ -9,11 +9,11 @@ import android.widget.TextView;
 
 import petstone.project.animalisland.R;
 
-public class MypageEventDialog {
+public class MypageNewsDialog {
     private Context context;
     String article;
 
-    public MypageEventDialog(Context context) {
+    public MypageNewsDialog(Context context) {
         this.context = context;
     }
 
@@ -25,11 +25,11 @@ public class MypageEventDialog {
         final Dialog dlg = new Dialog(context);
 
         dlg.requestWindowFeature(Window.FEATURE_NO_TITLE);
-        dlg.setContentView(R.layout.mypage_event_component);
+        dlg.setContentView(R.layout.mypage_news_component);
         dlg.show();
-        TextView article_tv = dlg.findViewById(R.id.mypage_event_article);
+        TextView article_tv = dlg.findViewById(R.id.mypage_news_article);
         article_tv.setText(article);
-        Button close = dlg.findViewById(R.id.mypage_event_button);
+        Button close = dlg.findViewById(R.id.mypage_news_button);
 
         close.setOnClickListener(new View.OnClickListener() {
             @Override
